+++ Calculator/Calculator.Logic/Model/EvaluatingExpressionVisitor.cs
@@ -3,18 +3,18 @@ using Calculator.Model;
 
 namespace Calculator.Logic.Model
 {
-    public class EvaluatingExpressionVisitor : AnExpressionVisitorWithResult<EvaluatingExpressionVisitor, double>
+    public class EvaluatingExpressionVisitor : AnExpressionVisitorWithResult<EvaluatingExpressionVisitor, decimal>
     {
-        public static double Evaluate(IExpression expression) => GetResultFor(expression);
-        protected override double UseVariable(string variable)
+        public static decimal Evaluate(IExpression expression) => GetResultFor(expression);
+        protected override decimal UseVariable(string variable)
         {
             throw new InvalidOperationException();
         }
-        protected override double UseParenthesed(double wrapped) => wrapped;
-        protected override double UseSubtraction(double left, double right) => left - right;
-        protected override double UseMultiplication(double left, double right) => left * right;
-        protected override double UseAddition(double left, double right) => left + right;
-        protected override double UseDivision(double left, double right) => left / right;
-        protected override double UseConstant(double value) => value;
+        protected override decimal UseParenthesed(decimal wrapped) => wrapped;
+        protected override decimal UseSubtraction(decimal left, decimal right) => left - right;
+        protected override decimal UseMultiplication(decimal left, decimal right) => left * right;
+        protected override decimal UseAddition(decimal left, decimal right) => left + right;
+        protected override decimal UseDivision(decimal left, decimal right) => left / right;
+        protected override decimal UseConstant(decimal value) => value;
     }
 }
