+++ Calculator/Calculator.Logic/Model/FormattingExpressionVisitor.cs
@@ -13,6 +13,6 @@ namespace Calculator.Logic.Model
         protected override string UseMultiplication(string left, string right) => $"{left}*{right}";
         protected override string UseAddition(string left, string right) => $"{left} + {right}";
         protected override string UseDivision(string left, string right) => $"{left}/{right}";
-        protected override string UseConstant(double value) => value.ToString(CultureInfo.InvariantCulture);
+        protected override string UseConstant(decimal value) => value.ToString(CultureInfo.InvariantCulture);
     }
 }
